backend-python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.1
openpyxl==3.1.2
PyYAML==6.0.1
Werkzeug==3.0.0
pydantic==2.4.2
pytest==7.4.3
python-multipart==0.0.6
requests==2.31.0
PyPDF2==3.0.1
pillow-heif==0.14.0
webp==0.1.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
scikit-image==0.21.0
